Update HybridBisectionSecantMethodForMultipleRoots.cpp
FInal c++ version of the hybrid bisection-secant method, now the error are polished and I was able to do the secant method inside the  the do while loop of the bisection method taking. This also have conditions to check divisions by zero and saves the and roots and the relative error of the secant method. I discovered this approach will use less iterations than the the hybrid newthon raphson method.